+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/meter/MeterSystem.java
@@ -189,18 +189,17 @@ public class MeterSystem implements Service {
         final String className = formatName(metricsName);
 
         /**
-         * Check is already define the metric
+         * Check whether the metrics class is already defined or not
          */
         try {
             CtClass alreadyMetric = classPool.get(METER_CLASS_PACKAGE + className);
             if (alreadyMetric.getSuperclass() != parentClass || type != meterPrototypes.get(metricsName).getScopeType()) {
-                throw new IllegalArgumentException("Already define metric " + metricsName
-                                                       + ", but calculate function or scope type is not same.");
+                throw new IllegalArgumentException("Already defined metrics class " + metricsName
+                                                       + ", but calculate function or scope type is different.");
             }
             log.debug("Metric {} is already define, so skip the metric creation.", metricsName);
             return true;
         } catch (NotFoundException e) {
-            // Ignore not found
         }
 
         CtClass metricsClass = classPool.makeClass(METER_CLASS_PACKAGE + className, parentClass);