+++ CHANGES.md
@@ -20,7 +20,7 @@ Release Notes.
 * Update the `timestamp` field type for `LogQuery`.
 * Support Zabbix protocol to receive agent metrics.
 * Update the Apdex metric combine calculator.
-* Enhancement `MeterSystem` allow creating metrics with same `metricName` / `function` / `scope`.
+* Enhance `MeterSystem` to allow creating metrics with same `metricName` / `function` / `scope`.
 
 #### UI
 * Update selector scroller to show in all pages.